+++ ceammc/CHANGELOG.md
@@ -14,3 +14,6 @@
 - array.copy added
 - array.do added
 - list.slice added
+
+### Changed:
+- pass.if renamed to pass_if